+++ test/perf/all_gatherv.cu
@@ -41,7 +41,7 @@ testResult_t AllGathervInitData(struct threadArgs* args, ncclDataType_t type, nc
   return testSuccess;
 }
 
-void AllGathervGetBw(size_t count, int typesize, double sec, double* algBw, double* busBw, int nranks) {
+void AllGathervGetBw(size_t count, size_t typesize, double sec, double* algBw, double* busBw, int nranks) {
   double baseBw = (double)(count * typesize * (nranks - 1)) / 1.0E9 / sec;
 #ifdef TRIANGULAR
   const double halfSize = (((double)nranks*nranks+1)/2) / (nranks*nranks);
