+++ test/perf/gather.cu
@@ -40,7 +40,7 @@ testResult_t GatherInitData(struct threadArgs* args, ncclDataType_t type, ncclRe
   return testSuccess;
 }
 
-void GatherGetBw(size_t count, int typesize, double sec, double* algBw, double* busBw, int nranks) {
+void GatherGetBw(size_t count, size_t typesize, double sec, double* algBw, double* busBw, int nranks) {
   double baseBw = (double)(count * nranks * typesize) / 1.0E9 / sec;
 
   *algBw = baseBw;
